Check If Exists Added.

+++ CodesceneReeinventTest/CodesceneReeinventTest/Application/Services/FileDownloader/FileDownloader.cs
@@ -15,10 +15,13 @@ namespace CodesceneReeinventTest.Application.Services.FileDownloader
         {
             try
             {
-                await DownloadAsync();
-                UnzipFile();
-                RenameFile();
-                DeleteFile();
+                if (!File.Exists(_artifactInfo.AbsoluteBinaryPath))
+                {
+                    await DownloadAsync();
+                    UnzipFile();
+                    RenameFile();
+                    DeleteFile();
+                }
             }
             catch (Exception ex)
             {
